+++ packages/sit-onyx/src/components/OnyxButton/OnyxButton.ct.tsx
@@ -1,6 +1,5 @@
-import happyIcon from "@sit-onyx/icons/emoji-happy-2.svg?raw";
 import { expect, test } from "../../playwright-axe";
-import { createScreenshotsForAllStates } from "../../utils/playwright";
+import { createScreenshotsForAllStates, mockPlaywrightIcon } from "../../utils/playwright";
 import OnyxButton from "./OnyxButton.vue";
 
 test("should render", async ({ mount, makeAxeBuilder }) => {
@@ -35,7 +34,7 @@ test("should display correctly when disabled", async ({ mount, makeAxeBuilder })
 test("should render button with icon", async ({ mount, makeAxeBuilder }) => {
   // ARRANGE
   const component = await mount(
-    <OnyxButton label="Button" variation="secondary" icon={happyIcon} />,
+    <OnyxButton label="Button" variation="secondary" icon={mockPlaywrightIcon} />,
   );
 
   // ASSERT
@@ -80,7 +79,7 @@ test(
           variation={variation}
           mode={mode}
           disabled={state === "disabled"}
-          icon={state === "icon" ? happyIcon : undefined}
+          icon={state === "icon" ? mockPlaywrightIcon : undefined}
         />,
       );