+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,18 @@
 # sit-onyx
 
+## 1.0.0-alpha.25
+
+### Minor Changes
+
+- 743d6b9: refactor(OnyxLoadingIndicator): update dots animation
+
+## 1.0.0-alpha.24
+
+### Patch Changes
+
+- Updated dependencies [5ac259d]
+  - @sit-onyx/icons@0.1.0-alpha.0
+
 ## 1.0.0-alpha.23
 
 ### Minor Changes